added tests for button/modal clicks
